entering and exiting parking lot +logs for entrances to the parking lot
added logs that log when the car enters the parking lot , plus checked insertion, extraction of cars from the parking lot , whether or not the insertion and extraction stop when we are getting to the top .
last but not least in this implementaion spaces that are faulty or saved by the workers or the robot , are done as per request

+++ client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/ComplaintController.java
@@ -58,7 +58,7 @@ public class ComplaintController {
     private ComboBox<String> orderSubscriptionBox;
 
     @FXML
-    private ComboBox<Integer> parkingLot;
+    private ComboBox<Long> parkingLot;
 
     @FXML
     private Button submitComplaint;
@@ -159,7 +159,7 @@ public class ComplaintController {
                 firstName.setText(user.getFirstName());
                 LastName.setText(user.getLastName());
                 email.setText(user.getEmail());
-                customerID.setText(Integer.toString(user.getId()));
+                customerID.setText(Long.toString(user.getId()));
 
                 message.setMessage("#getAllOrders");
                 SimpleClient.getClient().sendToServer(message);
